requirements_new.txt
aio_pika==9.5.5
APScheduler==3.11.0
av==12.3.0
beautifulsoup4==4.13.4
celery==5.5.3
fastapi==0.115.12
Flask==3.1.1
Flask_And_Redis==1.0.0
Flask_Bcrypt==1.0.1
flask_cors==5.0.1
Flask_JWT_Extended==4.6.0
flask_restx==1.3.0
Flask_SocketIO==5.5.1
flask_sqlalchemy==3.1.1
gevent==24.2.1
google_api_python_client==2.162.0
gTTS==2.5.4
jsonschema==4.24.0
moviepy==1.0.3
numpy
openai==1.88.0
opencv_python==4.11.0.86
paddleocr==3.0.1
pandas==2.0.3
Pillow==11.2.1
pillow_avif_plugin==1.5.2
pydub==0.25.1
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
pytz==2025.2
Requests==2.32.4
selenium==4.33.0
Shapely
solotop_random_user_agent==0.69.1
SQLAlchemy==2.0.41
srt==3.5.3
torch==2.7.1
tqdm==4.67.1
ultralytics==8.3.155
webdriver_manager==4.0.2
Werkzeug==3.1.3
